Checkpoint

+++ src/IceRpc/Transports/Internal/LogSlicFrameReaderDecorator.cs
@@ -83,7 +83,7 @@ namespace IceRpc.Transports.Internal
                 case FrameType.StreamReset:
                 {
                     StreamResetBody body = ReadFrame(() => reader.ReadStreamResetAsync(dataSize, default));
-                    _logger.LogReceivedSlicResetFrame(dataSize, (byte)body.ApplicationProtocolErrorCode);
+                    _logger.LogReceivedSlicResetFrame(dataSize, body.ApplicationProtocolErrorCode);
                     break;
                 }
                 case FrameType.StreamConsumed:
@@ -95,7 +95,7 @@ namespace IceRpc.Transports.Internal
                 case FrameType.StreamStopSending:
                 {
                     StreamStopSendingBody body = ReadFrame(() => reader.ReadStreamStopSendingAsync(dataSize, default));
-                    _logger.LogReceivedSlicStopSendingFrame(dataSize, (byte)body.ApplicationProtocolErrorCode);
+                    _logger.LogReceivedSlicStopSendingFrame(dataSize, body.ApplicationProtocolErrorCode);
                     break;
                 }
                 case FrameType.UnidirectionalStreamReleased:
